estfor/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud Firestore
google-cloud-firestore==2.13.1
google-auth==2.23.4

# HTTP client for EstFor API
httpx==0.25.2
requests==2.31.0

# Background task processing
celery==5.3.4
redis==5.0.1

# Monitoring and metrics
prometheus-client==0.19.0
structlog==23.2.0

# Data processing
pandas==2.1.4
numpy==1.25.2

# Configuration and validation
python-dotenv==1.0.0
marshmallow==3.20.1

# Security
cryptography==41.0.8
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9

# CORS
python-multipart==0.0.6

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3

# JSON handling
orjson==3.9.10

# Environment management
python-decouple==3.8

# Health checks
healthcheck==1.3.3
